% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotacoes_sequenciais_periodo.R
\name{cotacoes_sequenciais_periodo}
\alias{cotacoes_sequenciais_periodo}
\title{Obtém cotações sequenciais de uma moeda para um determinado período}
\usage{
cotacoes_sequenciais_periodo(moeda = "EUR-BRL", numero_resultados = 15)
}
\arguments{
\item{moeda}{Uma string que indica o par de moedas de interesse, no formato
"MOEDA1-MOEDA2". Por exemplo, "EUR-BRL" para o par Euro e Real brasileiro.
O valor padrão é "EUR-BRL".}

\item{numero_resultados}{Um inteiro que indica o número de resultados a ser
retornado pela API. O valor padrão é 15.}
}
\value{
Um tibble contendo as cotações da moeda para os períodos solicitados.
}
\description{
Esta função consulta uma API de cotações econômicas para obter uma sequência
de cotações para a moeda especificada. Os dados são retornados como um
tibble onde cada linha representa uma cotação em um momento específico.
}
\examples{
cotacoes_sequenciais_periodo()  # Retorna as últimas 15 cotações para EUR-BRL
cotacoes_sequenciais_periodo("USD-BRL", 10)  # Retorna as últimas 10 cotações para USD-BRL
}
